requirements.txt
whisper>=1.2.0
deep-translator>=1.7.0
pysrt>=1.1.2
python-docx>=0.8.11
PySide6>=6.2.0
torch